third_party/WebKit/LayoutTests/fast/css/import-with-bad-string.html

<!DOCTYPE html>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<style>
  @import url("bad-string-token
  );
  @import url();
</style>

<script>
test(() => {
    var rules = document.styleSheets[0].cssRules;
    assert_equals(rules.length, 1);
    assert_equals(rules[0].cssText, '@import url("");');
}, "@import rules with <bad-string-token> are invalid");
</script>
